ADD: Move evaluation to eval.py
